requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
pandas==2.1.0
openpyxl==3.1.2
Werkzeug==2.3.7
python-dotenv==1.0.0
reportlab==4.0.5
fpdf==1.7.2
PyJWT==2.8.0

